+++ lib/shared_map.dart
@@ -4,4 +4,4 @@ library shared_map;
 export 'src/shared_map_base.dart';
 export 'src/shared_map_cached.dart';
 export 'src/shared_map_extension.dart';
-export 'src/shared_object.dart';
+export 'src/shared_reference.dart';